requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
pytesseract
pillow
scikit-learn
joblib
python-multipart
httpx
pytest